+++ src/util/StrShuffler.js
@@ -11,56 +11,67 @@ for (let i = 32; i <= 126; i++) {
 */
 
 const mod = (n, m) => ((n % m) + m) % m;
-class StrShuffler {
-    static baseDictionary = '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz~-';
-    static shuffledIndicator = '_rhs';
-    static generateDictionary() {
-        let str = '';
-        const split = StrShuffler.baseDictionary.split('');
-        while (split.length > 0) {
-            str += split.splice(Math.floor(Math.random() * split.length), 1)[0];
-        }
-        return str;
+const baseDictionary = '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz~-';
+const shuffledIndicator = '_rhs';
+const generateDictionary = function () {
+    let str = '';
+    const split = baseDictionary.split('');
+    while (split.length > 0) {
+        str += split.splice(Math.floor(Math.random() * split.length), 1)[0];
     }
-
-    constructor(dictionary = StrShuffler.generateDictionary()) {
+    return str;
+};
+class StrShuffler {
+    constructor(dictionary = generateDictionary()) {
         this.dictionary = dictionary;
     }
     shuffle(str) {
-        if (str.startsWith(StrShuffler.shuffledIndicator)) {
+        if (str.startsWith(shuffledIndicator)) {
             return str;
         }
         let shuffledStr = '';
         for (let i = 0; i < str.length; i++) {
             const char = str.charAt(i);
-            const idx = StrShuffler.baseDictionary.indexOf(char);
-            if (idx === -1) {
+            const idx = baseDictionary.indexOf(char);
+            if (char === '%' && str.length - i >= 3) {
+                shuffledStr += char;
+                shuffledStr += str.charAt(++i);
+                shuffledStr += str.charAt(++i);
+            } else if (idx === -1) {
                 shuffledStr += char;
             } else {
-                shuffledStr += this.dictionary.charAt(mod(idx + i, StrShuffler.baseDictionary.length));
+                shuffledStr += this.dictionary.charAt(mod(idx + i, baseDictionary.length));
             }
         }
-        return StrShuffler.shuffledIndicator + shuffledStr;
+        return shuffledIndicator + shuffledStr;
     }
     unshuffle(str) {
-        if (!str.startsWith(StrShuffler.shuffledIndicator)) {
+        if (!str.startsWith(shuffledIndicator)) {
             return str;
         }
 
-        str = str.slice(StrShuffler.shuffledIndicator.length);
+        str = str.slice(shuffledIndicator.length);
 
         let unshuffledStr = '';
         for (let i = 0; i < str.length; i++) {
             const char = str.charAt(i);
             const idx = this.dictionary.indexOf(char);
-            if (idx === -1) {
+            if (char === '%' && str.length - i >= 3) {
+                unshuffledStr += char;
+                unshuffledStr += str.charAt(++i);
+                unshuffledStr += str.charAt(++i);
+            } else if (idx === -1) {
                 unshuffledStr += char;
             } else {
-                unshuffledStr += StrShuffler.baseDictionary.charAt(mod(idx - i, StrShuffler.baseDictionary.length));
+                unshuffledStr += baseDictionary.charAt(mod(idx - i, baseDictionary.length));
             }
         }
         return unshuffledStr;
     }
 }
+
+StrShuffler.baseDictionary = baseDictionary;
+StrShuffler.shuffledIndicator = shuffledIndicator;
+StrShuffler.generateDictionary = generateDictionary;
 
 module.exports = StrShuffler;